+++ packages/eslint-plugin/src/rules/index.ts
@@ -16,6 +16,8 @@ import strictExportDeclareModifiers = require("./strict-export-declare-modifiers
 import noSingleDeclareModule = require("./no-single-declare-module");
 import noOldDTHeader = require("./no-old-dt-header");
 import noImportOfDevDependencies = require("./no-import-of-dev-dependencies");
+import npmNaming = require("./npm-naming");
+import expect = require("./expect");
 
 export const rules = {
   "export-just-namespace": exportJustNamespace,
@@ -36,4 +38,6 @@ export const rules = {
   "no-single-declare-module": noSingleDeclareModule,
   "no-old-dt-header": noOldDTHeader,
   "no-import-of-dev-dependencies": noImportOfDevDependencies,
+  "npm-naming": npmNaming,
+  expect,
 };