+++ packages/eslint-plugin/src/util.ts
@@ -35,7 +35,7 @@ export function commentsMatching(
   }
 }
 
-function findUp<T extends {}>(p: string, fn: (p: string) => T | undefined): T | undefined {
+export function findUp<T extends {}>(p: string, fn: (p: string) => T | undefined): T | undefined {
   p = path.resolve(p);
   const root = path.parse(p).root;
 
@@ -115,4 +115,25 @@ export function getImportSource(
   }
 
   return undefined;
+}
+
+export function isMainFile(fileName: string, allowNested: boolean) {
+  // Linter may be run with cwd of the package. We want `index.d.ts` but not `submodule/index.d.ts` to match.
+  if (fileName === "index.d.ts") {
+    return true;
+  }
+
+  if (path.basename(fileName) !== "index.d.ts") {
+    return false;
+  }
+
+  let parent = path.dirname(fileName);
+  // May be a directory for an older version, e.g. `v0`.
+  // Note a types redirect `foo/ts3.1` should not have its own header.
+  if (allowNested && /^v(0\.)?\d+$/.test(path.basename(parent))) {
+    parent = path.dirname(parent);
+  }
+
+  // Allow "types/foo/index.d.ts", not "types/foo/utils/index.d.ts"
+  return path.basename(path.dirname(parent)) === "types";
 }